backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
anthropic>=0.40.0

# Supabase Authentication
supabase==2.11.0
PyJWT==2.9.0